refactor(workflow): Moved persistense directory into infrastructure
