+++ module/workflow/src/Infrastructure/Persistence/Projector/DbalStatusNameChangedEventProjector.php
@@ -2,20 +2,20 @@
 
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE.txt for license details.
+ * See license.txt for license details.
  */
 
 declare(strict_types = 1);
 
-namespace Ergonode\Workflow\Persistence\Dbal\Projector;
+namespace Ergonode\Workflow\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
-use Ergonode\Workflow\Domain\Event\Status\StatusDescriptionChangedEvent;
+use Ergonode\Workflow\Domain\Event\Status\StatusNameChangedEvent;
 use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class StatusDescriptionChangedEventProjector
+class DbalStatusNameChangedEventProjector
 {
     private const TABLE = 'status';
 
@@ -42,12 +42,12 @@ class StatusDescriptionChangedEventProjector
     /**
      * {@inheritDoc}
      */
-    public function __invoke(StatusDescriptionChangedEvent $event): void
+    public function __invoke(StatusNameChangedEvent $event): void
     {
         $this->connection->update(
             self::TABLE,
             [
-                'description' => $this->serializer->serialize($event->getTo()->getTranslations(), 'json'),
+                'name' => $this->serializer->serialize($event->getTo()->getTranslations(), 'json'),
             ],
             [
                 'id' => $event->getAggregateId()->getValue(),
