+++ module/workflow/src/Infrastructure/Persistence/Projector/Workflow/DbalWorkflowCreatedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Workflow\Persistence\Dbal\Projector\Workflow;
+namespace Ergonode\Workflow\Infrastructure\Persistence\Projector\Workflow;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -16,7 +16,7 @@ use Ergonode\Workflow\Domain\Event\Workflow\WorkflowCreatedEvent;
 
 /**
  */
-class WorkflowCreatedEventProjector
+class DbalWorkflowCreatedEventProjector
 {
     private const TABLE = 'workflow';
