+++ module/workflow/src/Infrastructure/Persistence/Projector/Workflow/DbalWorkflowTransitionRemovedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Workflow\Persistence\Dbal\Projector\Workflow;
+namespace Ergonode\Workflow\Infrastructure\Persistence\Projector\Workflow;
 
 use Doctrine\DBAL\Connection;
 use Ergonode\SharedKernel\Domain\Aggregate\StatusId;
@@ -15,7 +15,7 @@ use Ergonode\Workflow\Domain\Event\Workflow\WorkflowTransitionRemovedEvent;
 
 /**
  */
-class WorkflowTransitionRemovedEventProjector
+class DbalWorkflowTransitionRemovedEventProjector
 {
     private const TABLE = 'workflow_transition';
